chore: update README, move hash adapter to hash_adapter before it is properly fixed
