+++ lib/lutaml/model/config.rb
@@ -71,6 +71,9 @@ module Lutaml
       end
 
       def class_for(adapter, type)
+        # TODO: Remove once HashAdapter namespace is renamed to Hash
+        adapter = "hash_adapter" if adapter == "hash"
+
         Lutaml::Model.const_get(to_class_name(adapter))
           .const_get(to_class_name(type))
       end
@@ -86,6 +89,9 @@ module Lutaml
       end
 
       def load_adapter_file(adapter, type)
+        # TODO: Remove once HashAdapter namespace is renamed to Hash
+        adapter = "hash_adapter" if adapter == "hash"
+
         adapter_file = File.join(adapter, type)
         require_relative adapter_file
       rescue LoadError
