+++ lib/lutaml/model/hash_adapter.rb
@@ -7,11 +7,11 @@ module Lutaml
   end
 end
 
-require_relative "hash/standard_adapter"
-require_relative "hash/document"
-require_relative "hash/mapping"
-require_relative "hash/mapping_rule"
-require_relative "hash/transform"
+require_relative "hash_adapter/standard_adapter"
+require_relative "hash_adapter/document"
+require_relative "hash_adapter/mapping"
+require_relative "hash_adapter/mapping_rule"
+require_relative "hash_adapter/transform"
 
 Lutaml::Model::FormatRegistry.register(
   :hash,
