+++ lib/lutaml/model/hash_adapter/document.rb
