+++ lib/lutaml/model/hash_adapter/mapping.rb
