+++ lib/lutaml/model/hash_adapter/mapping_rule.rb
