+++ lib/lutaml/model/hash_adapter/standard_adapter.rb
