+++ lib/lutaml/model/hash_adapter/transform.rb
