+++ spec/spec_helper.rb
@@ -28,6 +28,7 @@ require_relative "../lib/lutaml/model"
 
 Lutaml::Model::Config.configure do |config|
   config.xml_adapter_type = :nokogiri
+  config.hash_adapter_type = :standard_hash
   config.json_adapter_type = :standard_json
   config.yaml_adapter_type = :standard_yaml
   config.toml_adapter_type = :toml_rb
